server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]",
  "groq>=1.0.0",
  "python-dotenv>=1.2.1",
  "uvicorn[standard]",
]
